Remove admin user,password pair from grafana configmap
The latest version of the Grafana helm chart doesn't allow the admin
user and password to be set in a configmap because it's readable by all
clients. This change removes those. The *secrets* created in
grafana/Tiltfile ought to be enough. If they're not, users can use
environment variables to pass the user and password to the binary.