feat: add apiserver clients certs to the grafana image
In order for the grafana-embedded apiserver to aggregate to
the apiserver running in the kind cluster the apiserver
needs to present the kind-apiserver with certs as part of
the auth handshake.

These certs are copied from grafana-enterprise and checked in here.

The other half of this will be in ops-devenv, which will have the ca
so the kind-apiserver can validate the certs